+++ extensions/easydict/src/detectLanauge/utils.ts
@@ -2,26 +2,27 @@
  * @author: tisfeng
  * @createTime: 2022-08-12 18:37
  * @lastEditor: tisfeng
- * @lastEditTime: 2022-09-12 10:34
+ * @lastEditTime: 2022-09-25 22:56
  * @fileName: utils.ts
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
  */
 
 import { preferredLanguages } from "../preferences";
+import { englishLanguageItem } from "../language/consts";
 
 /**
  * check if the language is preferred language
  */
 export function isPreferredLanguage(languageId: string): boolean {
-  return preferredLanguages.map((item) => item.youdaoId).includes(languageId);
+  return preferredLanguages.map((item) => item.youdaoLangCode).includes(languageId);
 }
 
 /**
  * check if preferred languages contains English language
  */
 export function checkIfPreferredLanguagesContainEnglish(): boolean {
-  return preferredLanguages.find((item) => item.youdaoId === "en") !== undefined;
+  return preferredLanguages.find((item) => item.youdaoLangCode === englishLanguageItem.youdaoLangCode) !== undefined;
 }
 
 /**
@@ -29,7 +30,7 @@ export function checkIfPreferredLanguagesContainEnglish(): boolean {
  */
 export function checkIfPreferredLanguagesContainChinese(): boolean {
   const lanuguageIdPrefix = "zh";
-  return preferredLanguages.find((item) => item.youdaoId.startsWith(lanuguageIdPrefix)) !== undefined;
+  return preferredLanguages.find((item) => item.youdaoLangCode.startsWith(lanuguageIdPrefix)) !== undefined;
 }
 
 /**